server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["RChandler234 <29521172+RChandler234@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
waitress = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
pg8000 = "^1.31.2"
sqlalchemy = "^2.0.36"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
flask-cors = "^5.0.0"
flasgger = "^0.9.7.1"
flask-restful = "^0.3.10"
python-dotenv = "^1.0.1"
flask-testing = "^0.8.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.3"
black = "^24.10.0"

[tool.flake8]
ignore = ["F401","E402"]
exclude = [".venv"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
"start" = "app:run_app"
"debug" = "app:run_app_debug"

